Move meeting to public and update contract
