+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './private/index';
 export * from './public/index';
+export { meeting } from './public/meeting';
