+++ src/private/index.ts
@@ -26,4 +26,3 @@ export {
 } from './privateAPIs';
 export { conversations } from './conversations';
 export { meetingRoom } from './meetingRoom';
-export { meeting } from './meeting';
